refactor(simulation): group feature use cases and results
