+++ src/main/java/com/renewsim/backend/simulation_service/create/application/port/in/CreateRealSimulationUseCase.java
@@ -1,0 +1,8 @@
+package com.renewsim.backend.simulation_service.create.application.port.in;
+
+import com.renewsim.backend.simulation_service.shared.application.SimulationDetailsResult;
+import com.renewsim.backend.simulation_service.create.application.command.CreateRealSimulationCommand;
+
+public interface CreateRealSimulationUseCase {
+    SimulationDetailsResult createSimulation(CreateRealSimulationCommand command);
+}
