+++ src/main/java/com/renewsim/backend/simulation_service/delete/application/port/in/DeleteRealSimulationUseCase.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.delete.application;
+package com.renewsim.backend.simulation_service.delete.application.port.in;
 
 public interface DeleteRealSimulationUseCase {
     void deleteSimulation(Long id, String requesterUsername, boolean isAdmin);
