--- src/main/java/com/renewsim/backend/simulation_service/detail/application/GetRealSimulationUseCase.java
@@ -1,7 +1,0 @@
-package com.renewsim.backend.simulation_service.detail.application;
-
-import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
-
-public interface GetRealSimulationUseCase {
-    SimulationDetailsResult getSimulationById(Long id, String requesterUsername, boolean isAdmin);
-}
