--- src/main/java/com/renewsim/backend/simulation_service/history/application/ListUserRealSimulationsUseCase.java
@@ -1,5 +1,0 @@
-package com.renewsim.backend.simulation_service.history.application;
-
-public interface ListUserRealSimulationsUseCase {
-    UserSimulationListResult getUserSimulations(String username);
-}
