+++ src/main/java/com/renewsim/backend/simulation_service/history/application/port/in/ListUserRealSimulationsUseCase.java
@@ -1,0 +1,7 @@
+package com.renewsim.backend.simulation_service.history.application.port.in;
+
+import com.renewsim.backend.simulation_service.history.application.result.UserSimulationListResult;
+
+public interface ListUserRealSimulationsUseCase {
+    UserSimulationListResult getUserSimulations(String username);
+}
